refactor: converted react queries into their various custom hooks
