+++ src/features/cabins/CabinTable.jsx
@@ -1,5 +1,4 @@
-import { useQuery } from "@tanstack/react-query";
-import { getCabins } from "../../services/apiCabins";
+import { useCabins } from "./useCabins";
 
 import Spinner from "../../ui/Spinner";
 import CabinRow from "./CabinRow";
@@ -31,14 +30,7 @@ const TableHeader = styled.header`
 `;
 
 function CabinTable() {
-  const {
-    isLoading,
-    data: cabins,
-    error,
-  } = useQuery({
-    queryKey: ["cabins"],
-    queryFn: getCabins,
-  });
+  const { isLoading, cabins } = useCabins();
 
   console.log(cabins);
